Spectre.Interactive

Description

Interactive Communication between Haskell and Spectre

Documentation

data Session

Spectre Interactive Session

Constructors

Session

Fields

pty :: !Pty

dir :: !FilePath

type Parameter = String

Name of a netlist Parameter

Session Management

startSession :: [FilePath] -> FilePath -> FilePath -> IO Session

Initialize spectre session with given include path, netlist and temp dir

startSession' :: [FilePath] -> FilePath -> IO Session

Initialize spectre session with given include path and netlist

stopSession :: Session -> IO ()

Close a spectre interactive session

Running Simulations

runAll :: Session -> IO NutMeg

Run all simulation analyses

listAnalysis :: Session -> IO [(String, Analysis)]

Get Map of Available Simulation Analyses: (id, type)

runAnalysis :: Session -> String -> IO NutMeg

Run Selected Analysis only

Netlist Parameters

getParameter :: Session -> Parameter -> IO Double

Get Netlist Parameter

setParameter :: Session -> Parameter -> Double -> IO ()

Set Netlist Parameter

getParameters :: Session -> [Parameter] -> IO (Map Parameter Double)

Get a list of Parameters as Map

setParameters :: Session -> Map Parameter Double -> IO (Map Parameter ())

Get a list of Parameters